+++ src/Common/ExprCalc.Entities/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace ExprCalc.Entities
-{
-    public class Class1
-    {
-
-    }
-}