+++ src/CoreLogic/ExprCalc.CoreLogic.Api/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace ExprCalc.CoreLogic.Api
-{
-    public class Class1
-    {
-
-    }
-}