+++ src/CoreLogic/ExprCalc.CoreLogic/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace ExprCalc.CoreLogic
-{
-    public class Class1
-    {
-
-    }
-}