+++ CHANGELOG.md
@@ -8,8 +8,9 @@ Expression calculator changelog
 - Basic CI added (#12)
 - Expression parsing (#13)
 - Projects for the main components of the system (#14)
-- Main project and Rest API projects initialized(#20)
+- Main project and Rest API projects initialized (#20)
 - Code coverage reports uploading to codecov.io (#20)
+- Core logic project intitialized (#21)
 
 #### Changed